refactor: javascript files now es6 with promises
