--- src/js/dbhelper.js
@@ -1,161 +1,0 @@
-import Config from '../../config';
-
-/**
- * Common database helper functions.
- */
-export default class DBHelper {
-    /**
-     * Database URL.
-     * Change this to restaurants.json file location on your server.
-     */
-    static get DATABASE_URL() {
-        return Config.DATA_STORE || 'http://localhost:8000/data/restaurants.json';
-    }
-
-    /**
-     * Fetch all restaurants.
-     */
-    static fetchRestaurants(callback) {
-        fetch(DBHelper.DATABASE_URL)
-            .then(response => response.json())
-            .then(json => callback(null, json.restaurants))
-            .catch(error => callback(error,null));
-    }
-
-    /**
-     * Fetch a restaurant by its ID.
-     */
-    static fetchRestaurantById(id, callback) {
-        // fetch all restaurants with proper error handling.
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                const restaurant = restaurants.find(r => r.id == id);
-                if (restaurant) { // Got the restaurant
-                    callback(null, restaurant);
-                } else { // Restaurant does not exist in the database
-                    callback('Restaurant does not exist', null);
-                }
-            }
-        });
-    }
-
-    /**
-     * Fetch restaurants by a cuisine type with proper error handling.
-     */
-    static fetchRestaurantByCuisine(cuisine, callback) {
-        // Fetch all restaurants  with proper error handling
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                // Filter restaurants to have only given cuisine type
-                const results = restaurants.filter(r => r.cuisine_type == cuisine);
-                callback(null, results);
-            }
-        });
-    }
-
-    /**
-     * Fetch restaurants by a neighborhood with proper error handling.
-     */
-    static fetchRestaurantByNeighborhood(neighborhood, callback) {
-        // Fetch all restaurants
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                // Filter restaurants to have only given neighborhood
-                const results = restaurants.filter(r => r.neighborhood == neighborhood);
-                callback(null, results);
-            }
-        });
-    }
-
-    /**
-     * Fetch restaurants by a cuisine and a neighborhood with proper error handling.
-     */
-    static fetchRestaurantByCuisineAndNeighborhood(cuisine, neighborhood, callback) {
-        // Fetch all restaurants
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                let results = restaurants;
-                if (cuisine !== 'all') { // filter by cuisine
-                    results = results.filter(r => r.cuisine_type == cuisine);
-                }
-                if (neighborhood !== 'all') { // filter by neighborhood
-                    results = results.filter(r => r.neighborhood == neighborhood);
-                }
-                callback(null, results);
-            }
-        });
-    }
-
-    /**
-     * Fetch all neighborhoods with proper error handling.
-     */
-    static fetchNeighborhoods(callback) {
-        // Fetch all restaurants
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                // Get all neighborhoods from all restaurants
-                const neighborhoods = restaurants.map((v, i) => restaurants[i].neighborhood);
-                // Remove duplicates from neighborhoods
-                const uniqueNeighborhoods = neighborhoods.filter((v, i) => neighborhoods.indexOf(v) == i);
-                callback(null, uniqueNeighborhoods);
-            }
-        });
-    }
-
-    /**
-     * Fetch all cuisines with proper error handling.
-     */
-    static fetchCuisines(callback) {
-        // Fetch all restaurants
-        DBHelper.fetchRestaurants((error, restaurants) => {
-            if (error) {
-                callback(error, null);
-            } else {
-                // Get all cuisines from all restaurants
-                const cuisines = restaurants.map((v, i) => restaurants[i].cuisine_type);
-                // Remove duplicates from cuisines
-                const uniqueCuisines = cuisines.filter((v, i) => cuisines.indexOf(v) == i);
-                callback(null, uniqueCuisines);
-            }
-        });
-    }
-
-    /**
-     * Restaurant page URL.
-     */
-    static urlForRestaurant(restaurant) {
-        return (`./restaurant.html?id=${restaurant.id}`);
-    }
-
-    /**
-     * Restaurant image URL.
-     */
-    static imageUrlForRestaurant(restaurant) {
-        return (`/img/${restaurant.photograph}`);
-    }
-
-    /**
-     * Map marker for a restaurant.
-     */
-    static mapMarkerForRestaurant(restaurant, map) {
-        return new window.google.maps.Marker({
-                position: restaurant.latlng,
-                title: restaurant.name,
-                url: DBHelper.urlForRestaurant(restaurant),
-                map: map,
-                animation: window.google.maps.Animation.DROP
-            }
-        );
-    }
-
-}
